ColorDialog
refactor
